fitrep legend as part of page, not rerendered repeatedly
