+++ src/js/graph/graph.js
@@ -48,9 +48,6 @@ export const draw_psr_viz = (data) => {
     // AT/CC Reporting Senior Bars
     make_atcc_reporting_senior_bars(data);
 
-    //Fitrep graph legend
-    draw_legend();
-
     draw_fitrep_graph(data, fitreps_g);
 };
 
@@ -219,199 +216,6 @@ function make_atcc_reporting_senior_bars(data) {
         rep_sen_at_g,
         lib.at_cc_command_bar_color
     );
-}
-
-function draw_legend() {
-    const { legend_canvas } = getPageElements();
-
-    // #region Add FITREP plot legend
-    // 2 groups in the legend canvas, 1 for the promotion recommendation legend, 1 for the traffic size legend
-    const prom_rec_g = legend_canvas
-        .append('g')
-        .attr('transform', `translate(10,${8 * lib.bar_height})`);
-    const traffic_g = legend_canvas
-        .append('g')
-        .attr('transform', `translate(100,${8 * lib.bar_height})`);
-
-    // Draw the promotion recommendation legend
-    /// Markers
-    const prom_rec_legend_marker_groups = prom_rec_g
-        .selectAll('g')
-        .data(lib.prom_rec_categories)
-        .enter()
-        .append('g')
-        .attr(
-            'transform',
-            (d, i) =>
-                `translate(0, ${
-                    i *
-                    1.2 *
-                    Math.sqrt(
-                        (4 *
-                            lib.fitrep_marker_size(
-                                lib.fitrep_legend_marker_size
-                            )) /
-                            Math.PI
-                    )
-                })`
-        );
-    /// Outlines
-    prom_rec_legend_marker_groups
-        .append('path')
-        .attr('d', function (d) {
-            var symbol = lib.fitrep_shape_scale(d.toUpperCase());
-            var size = lib.fitrep_marker_size(lib.fitrep_legend_marker_size);
-            return symbol.size(size)();
-        })
-        .attr('fill', (d) => lib.fitrep_color_scale(d.toUpperCase()))
-        .attr('opacity', lib.fitrep_marker_opacity);
-    /// Marker outlines
-    prom_rec_legend_marker_groups
-        .append('path')
-        .attr('d', function (d) {
-            var symbol = lib.fitrep_shape_scale(d.toUpperCase());
-            var size = lib.fitrep_marker_size(lib.fitrep_legend_marker_size);
-            return symbol.size(size)();
-        })
-        .attr('fill', 'none')
-        .attr('stroke', 'black')
-        .attr('stroke-width', lib.fitrep_marker_stroke_width);
-    /// Labels
-    prom_rec_legend_marker_groups
-        .append('text')
-        .attr(
-            'transform',
-            `translate(${
-                1.1 *
-                Math.sqrt(
-                    (4 *
-                        lib.fitrep_marker_size(lib.fitrep_legend_marker_size)) /
-                        Math.PI
-                )
-            })`
-        )
-        .text((d) => d);
-
-    // Draw the "traffic" legend
-    /// Markers
-    const fitrep_traffic_legend_marker_groups = traffic_g
-        .selectAll('g')
-        .data(lib.fitrep_traffic_legend_sizes)
-        .enter()
-        .append('g')
-        .attr(
-            'transform',
-            (d, i) =>
-                `translate(0, ${
-                    i *
-                    1.2 *
-                    Math.sqrt(
-                        (4 *
-                            lib.fitrep_marker_size(
-                                Math.max(...lib.fitrep_traffic_legend_sizes)
-                            )) /
-                            Math.PI
-                    )
-                })`
-        );
-    /// Markers
-    fitrep_traffic_legend_marker_groups
-        .append('path')
-        .attr('d', function (d) {
-            var symbol = d3.symbol(d3.symbolCircle);
-            var size = lib.fitrep_marker_size(d);
-            return symbol.size(size)();
-        })
-        .attr('fill', (d) => lib.fitrep_traffic_legend_marker_color)
-        .attr('opacity', lib.fitrep_marker_opacity);
-    /// Marker outlines
-    fitrep_traffic_legend_marker_groups
-        .append('path')
-        .attr('d', function (d) {
-            var symbol = d3.symbol(d3.symbolCircle);
-            var size = lib.fitrep_marker_size(d);
-            return symbol.size(size)();
-        })
-        .attr('fill', 'none')
-        .attr('stroke', 'black')
-        .attr('stroke-width', lib.fitrep_marker_stroke_width);
-    /// Labels
-    fitrep_traffic_legend_marker_groups
-        .append('text')
-        .attr(
-            'transform',
-            `translate(${
-                1.1 *
-                Math.sqrt(
-                    (4 *
-                        lib.fitrep_marker_size(lib.fitrep_legend_marker_size)) /
-                        Math.PI
-                )
-            })`
-        )
-        .text((d) => d);
-
-    // #endregion
-
-    const command_rg_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 1.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('Regular Commands')
-        .style('dominant-baseline', 'middle');
-
-    const rep_sen_rg_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 2.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('Regular Reporting Seniors')
-        .style('dominant-baseline', 'middle');
-
-    const command_idt_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 3.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('IDT Commands')
-        .style('dominant-baseline', 'middle');
-
-    const rep_sen_idt_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 4.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('IDT Reporting Seniors')
-        .style('dominant-baseline', 'middle');
-
-    const command_at_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 5.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('AT Commands')
-        .style('dominant-baseline', 'middle');
-
-    const rep_sen_at_label_g = legend_canvas
-        .append('g')
-        .attr(
-            'transform',
-            `translate(0, ${lib.margin.gap + 6.5 * lib.bar_height})`
-        )
-        .append('text')
-        .text('AT Reporting Seniors')
-        .style('dominant-baseline', 'middle');
 }
 
 function draw_axises(group, data) {
